task 6
Добавил автоматическую загрузку ленты новостей при ее добавлении.
Исправил работу счетчика непрочитанных новостей

+++ RSSReader/Classes/NewsItem/Model/NewsSourse.h
@@ -48,5 +48,5 @@
 
 - (void)cancelDownload;
 
-- (int) numberOfUnreadNews;
+- (NSInteger) numberOfUnreadNews;
 @end
